Another small fix to prevent annoying warning messages in the log.
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@784559 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/manufacturing/src/org/ofbiz/manufacturing/mrp/InventoryEventPlannedServices.java
@@ -73,7 +73,7 @@ public class InventoryEventPlannedServices {
             mrpEvent.create();
         } else {
             BigDecimal qties = newQuantity.add((BigDecimal)mrpEvent.getBigDecimal("quantity"));
-            mrpEvent.put("quantity", qties);
+            mrpEvent.put("quantity", qties.doubleValue());
             if (!UtilValidate.isEmpty(eventName)) {
                 String existingEventName = mrpEvent.getString("eventName");
                 mrpEvent.put("eventName", (UtilValidate.isEmpty(existingEventName)? eventName: existingEventName + ", " + eventName));
